Convert collection of html nodes into data frame

Convert collection of html nodes into data frame

html_attrs_dfr(x, attrs = NULL, trim = FALSE,
  defaults = NA_character_, add_text = TRUE)

Arguments

x
xml_nodeset object, containing text and attributes of interest

attrs
character vector of attribute names. If missing, all attributes will be used

trim
if TRUE, will trim leading and trailing spaces

defaults
character vector of default values to be passed to rvest::html_attr(). Recycled to match length of attrs

add_text
if TRUE, node content will be added as .text column (using rvest::html_text)

Value

data frame with one row per xml node, consisting of an html_text column with text and additional columns with attributes

Examples

library(polite)
library(rvest)

#> Loading required package: xml2

bow("https://en.wikipedia.org/wiki/List_of_cognitive_biases") %>%
  scrape() %>%
  html_nodes("tr td:nth-child(1) a") %>%
  html_attrs_dfr()

#>                                                      href
#> 1                                  /wiki/Ambiguity_effect
#> 2                                         /wiki/Anchoring
#> 3                          /wiki/Anthropocentric_thinking
#> 4   /wiki/Anthropomorphism#Psychology_of_Anthropomorphism
#> 5                                  /wiki/Attentional_bias
#> 6                                   /wiki/Automation_bias
#> 7                            /wiki/Availability_heuristic
#> 8                              /wiki/Availability_cascade
#> 9                                   /wiki/Backfire_effect
#> 10                                 /wiki/Bandwagon_effect
#> 11                                /wiki/Base_rate_fallacy
#> 12                                      /wiki/Belief_bias
#> 13                              /wiki/Ben_Franklin_effect
#> 14                              /wiki/Berkson%27s_paradox
#> 15                                  /wiki/Bias_blind_spot
#> 16                           /wiki/Choice-supportive_bias
#> 17                              /wiki/Clustering_illusion
#> 18                                /wiki/Confirmation_bias
#> 19                                  /wiki/Congruence_bias
#> 20                              /wiki/Conjunction_fallacy
#> 21                   /wiki/Conservatism_(belief_revision)
#> 22                       /wiki/Continued_influence_effect
#> 23                                  /wiki/Contrast_effect
#> 24                                  /wiki/Shy_Tory_Factor
#> 25                               /wiki/Curse_of_knowledge
#> 26                                        /wiki/Declinism
#> 27                                     /wiki/Decoy_effect
#> 28                                   /wiki/Default_effect
#> 29                              /wiki/Denomination_effect
#> 30                               /wiki/Disposition_effect
#> 31                                 /wiki/Distinction_bias
#> 32                    /wiki/Dunning%E2%80%93Kruger_effect
#> 33                                 /wiki/Duration_neglect
#> 34                                      /wiki/Empathy_gap
#> 35                                 /wiki/Endowment_effect
#> 36                            /wiki/Experimenter%27s_bias
#> 37                                 /wiki/Expectation_bias
#> 38                                  /wiki/Focusing_effect
#> 39                                     /wiki/Forer_effect
#> 40                                    /wiki/Barnum_effect
#> 41                      /wiki/Framing_effect_(psychology)
#> 42                    /wiki/Baader%E2%80%93Meinhof_effect
#> 43                             /wiki/Functional_fixedness
#> 44                              /wiki/Gambler%27s_fallacy
#> 45                                       /wiki/Groupthink
#> 46                         /wiki/Hard%E2%80%93easy_effect
#> 47                                   /wiki/Hindsight_bias
#> 48                         /wiki/Hostile_attribution_bias
#> 49                                 /wiki/Hot-hand_fallacy
#> 50                           /wiki/Hyperbolic_discounting
#> 51                       /wiki/Identifiable_victim_effect
#> 52                                      /wiki/IKEA_effect
#> 53                             /wiki/Illicit_transference
#> 54                              /wiki/Illusion_of_control
#> 55                             /wiki/Illusion_of_validity
#> 56                             /wiki/Illusory_correlation
#> 57                            /wiki/Illusory_truth_effect
#> 58                                      /wiki/Impact_bias
#> 59                        /wiki/Implicit_association_test
#> 60                    /wiki/Information_bias_(psychology)
#> 61                     /wiki/Insensitivity_to_sample_size
#> 62                            /wiki/Irrational_escalation
#> 63                            /wiki/Law_of_the_instrument
#> 64                            /wiki/Less-is-better_effect
#> 65                            /wiki/Look-elsewhere_effect
#> 66                                    /wiki/Loss_aversion
#> 67                             /wiki/Mere_exposure_effect
#> 68                                   /wiki/Money_illusion
#> 69                                 /wiki/Moral_credential
#> 70                                  /wiki/Negativity_bias
#> 71                           /wiki/Neglect_of_probability
#> 72                                    /wiki/Normalcy_bias
#> 73                                /wiki/Not_invented_here
#> 74                       /wiki/Observer-expectancy_effect
#> 75                                    /wiki/Omission_bias
#> 76                                    /wiki/Optimism_bias
#> 77                                   /wiki/Ostrich_effect
#> 78                                     /wiki/Outcome_bias
#> 79                            /wiki/Overconfidence_effect
#> 80                                       /wiki/Pareidolia
#> 81                                 /wiki/Pygmalion_effect
#> 82                                   /wiki/Pessimism_bias
#> 83                                 /wiki/Planning_fallacy
#> 84                    /wiki/Post-purchase_rationalization
#> 85                                     /wiki/Present_bias
#> 86                              /wiki/Pro-innovation_bias
#> 87                                  /wiki/Projection_bias
#> 88                           /wiki/Pseudocertainty_effect
#> 89                           /wiki/Reactance_(psychology)
#> 90                             /wiki/Reactive_devaluation
#> 91                                 /wiki/Recency_illusion
#> 92                                   /wiki/Restraint_bias
#> 93                           /wiki/Rhyme_as_reason_effect
#> 94                                /wiki/Risk_compensation
#> 95                                    /wiki/Salience_bias
#> 96                                   /wiki/Selection_bias
#> 97                             /wiki/Selective_perception
#> 98                                /wiki/Semmelweis_reflex
#> 99                       /wiki/Sexual_overperception_bias
#> 100                          /wiki/Social_comparison_bias
#> 101                        /wiki/Social_desirability_bias
#> 102                                 /wiki/Status_quo_bias
#> 103                                    /wiki/Stereotyping
#> 104                            /wiki/Subadditivity_effect
#> 105                           /wiki/Subjective_validation
#> 106                                     /wiki/Surrogation
#> 107                               /wiki/Survivorship_bias
#> 108                                /wiki/Time-saving_bias
#> 109                             /wiki/Third-person_effect
#> 110                 /wiki/Parkinson%27s_law_of_triviality
#> 111                       /wiki/Weber%E2%80%93Fechner_law
#> 112                      /wiki/Well_travelled_road_effect
#> 113                      /wiki/Women_are_wonderful_effect
#> 114                                  /wiki/Zero-risk_bias
#> 115                               /wiki/Zero-sum_thinking
#> 116                             /wiki/Actor-observer_bias
#> 117                                  /wiki/Authority_bias
#> 118                              /wiki/Cheerleader_effect
#> 119                /wiki/Defensive_attribution_hypothesis
#> 120                                 /wiki/Egocentric_bias
#> 121                       /wiki/Extrinsic_incentives_bias
#> 122                          /wiki/False_consensus_effect
#> 123                           /wiki/False_uniqueness_bias
#> 124                   /wiki/Fundamental_attribution_error
#> 125                         /wiki/Group_attribution_error
#> 126                                     /wiki/Halo_effect
#> 127                  /wiki/Illusion_of_asymmetric_insight
#> 128                     /wiki/Illusion_of_external_agency
#> 129                        /wiki/Illusion_of_transparency
#> 130                            /wiki/Illusory_superiority
#> 131                                    /wiki/Ingroup_bias
#> 132                           /wiki/Just-world_hypothesis
#> 133                                      /wiki/Moral_luck
#> 134                             /wiki/Na%C3%AFve_cynicism
#> 135                 /wiki/Na%C3%AFve_realism_(psychology)
#> 136                       /wiki/Outgroup_homogeneity_bias
#> 137                               /wiki/Self-serving_bias
#> 138                         /wiki/Shared_information_bias
#> 139                            /wiki/System_justification
#> 140                           /wiki/Trait_ascription_bias
#> 141                      /wiki/Ultimate_attribution_error
#> 142                       /wiki/Worse-than-average_effect
#> 143                              /wiki/Bizarreness_effect
#> 144                        /wiki/Cue-dependent_forgetting
#> 145                               /wiki/Cross-race_effect
#> 146                                    /wiki/Cryptomnesia
#> 147                                 /wiki/Egocentric_bias
#> 148                              /wiki/Fading_affect_bias
#> 149                                    /wiki/False_memory
#> 150                               /wiki/Generation_effect
#> 151                                   /wiki/Google_effect
#> 152                         /wiki/Leveling_and_sharpening
#> 153                     /wiki/Levels-of-processing_effect
#> 154                           /wiki/Misinformation_effect
#> 155                                 /wiki/Modality_effect
#> 156                        /wiki/Cue-dependent_forgetting
#> 157                         /wiki/Part-list_cueing_effect
#> 158                                   /wiki/Peak-end_rule
#> 159                      /wiki/Picture_superiority_effect
#> 160               /wiki/Socioemotional_selectivity_theory
#> 161                                  /wiki/Primacy_effect
#> 162                                  /wiki/Recency_effect
#> 163                          /wiki/Serial_position_effect
#> 164                               /wiki/Reminiscence_bump
#> 165                        /wiki/Misattribution_of_memory
#> 166                                  /wiki/Spacing_effect
#> 167                                /wiki/Spotlight_effect
#> 168                         /wiki/Suggestibility#External
#> 169                                    /wiki/Tachypsychia
#> 170                              /wiki/Telescoping_effect
#> 171                                  /wiki/Testing_effect
#> 172                               /wiki/Tip_of_the_tongue
#> 173                             /wiki/Von_Restorff_effect
#> 174                                /wiki/Zeigarnik_effect
#> 175                                                  <NA>
#> 176                           /wiki/List_of_memory_biases
#>                                 title                class
#> 1                    Ambiguity effect                 <NA>
#> 2                           Anchoring                 <NA>
#> 3            Anthropocentric thinking          mw-redirect
#> 4                    Anthropomorphism                 <NA>
#> 5                    Attentional bias                 <NA>
#> 6                     Automation bias                 <NA>
#> 7              Availability heuristic                 <NA>
#> 8                Availability cascade                 <NA>
#> 9                     Backfire effect          mw-redirect
#> 10                   Bandwagon effect                 <NA>
#> 11                  Base rate fallacy                 <NA>
#> 12                        Belief bias                 <NA>
#> 13                Ben Franklin effect                 <NA>
#> 14                  Berkson's paradox                 <NA>
#> 15                    Bias blind spot                 <NA>
#> 16             Choice-supportive bias                 <NA>
#> 17                Clustering illusion                 <NA>
#> 18                  Confirmation bias                 <NA>
#> 19                    Congruence bias                 <NA>
#> 20                Conjunction fallacy                 <NA>
#> 21     Conservatism (belief revision)                 <NA>
#> 22         Continued influence effect          mw-redirect
#> 23                    Contrast effect                 <NA>
#> 24                    Shy Tory Factor          mw-redirect
#> 25                 Curse of knowledge                 <NA>
#> 26                          Declinism                 <NA>
#> 27                       Decoy effect                 <NA>
#> 28                     Default effect                 <NA>
#> 29                Denomination effect                 <NA>
#> 30                 Disposition effect                 <NA>
#> 31                   Distinction bias                 <NA>
#> 32              Dunning–Kruger effect                 <NA>
#> 33                   Duration neglect                 <NA>
#> 34                        Empathy gap                 <NA>
#> 35                   Endowment effect                 <NA>
#> 36                Experimenter's bias          mw-redirect
#> 37                   Expectation bias          mw-redirect
#> 38                    Focusing effect          mw-redirect
#> 39                       Forer effect          mw-redirect
#> 40                      Barnum effect                 <NA>
#> 41        Framing effect (psychology)                 <NA>
#> 42              Baader–Meinhof effect                 <NA>
#> 43               Functional fixedness                 <NA>
#> 44                  Gambler's fallacy                 <NA>
#> 45                         Groupthink                 <NA>
#> 46                   Hard–easy effect                 <NA>
#> 47                     Hindsight bias                 <NA>
#> 48           Hostile attribution bias                 <NA>
#> 49                   Hot-hand fallacy          mw-redirect
#> 50             Hyperbolic discounting                 <NA>
#> 51         Identifiable victim effect                 <NA>
#> 52                        IKEA effect                 <NA>
#> 53               Illicit transference          mw-redirect
#> 54                Illusion of control                 <NA>
#> 55               Illusion of validity                 <NA>
#> 56               Illusory correlation                 <NA>
#> 57              Illusory truth effect                 <NA>
#> 58                        Impact bias                 <NA>
#> 59          Implicit association test          mw-redirect
#> 60      Information bias (psychology)                 <NA>
#> 61       Insensitivity to sample size                 <NA>
#> 62              Irrational escalation          mw-redirect
#> 63              Law of the instrument                 <NA>
#> 64              Less-is-better effect                 <NA>
#> 65              Look-elsewhere effect                 <NA>
#> 66                      Loss aversion                 <NA>
#> 67               Mere exposure effect          mw-redirect
#> 68                     Money illusion                 <NA>
#> 69                   Moral credential          mw-redirect
#> 70                    Negativity bias                 <NA>
#> 71             Neglect of probability                 <NA>
#> 72                      Normalcy bias                 <NA>
#> 73                  Not invented here                 <NA>
#> 74         Observer-expectancy effect                 <NA>
#> 75                      Omission bias                 <NA>
#> 76                      Optimism bias                 <NA>
#> 77                     Ostrich effect                 <NA>
#> 78                       Outcome bias                 <NA>
#> 79              Overconfidence effect                 <NA>
#> 80                         Pareidolia                 <NA>
#> 81                   Pygmalion effect                 <NA>
#> 82                     Pessimism bias          mw-redirect
#> 83                   Planning fallacy                 <NA>
#> 84      Post-purchase rationalization          mw-redirect
#> 85                       Present bias                 <NA>
#> 86                Pro-innovation bias                 <NA>
#> 87                    Projection bias          mw-redirect
#> 88             Pseudocertainty effect                 <NA>
#> 89             Reactance (psychology)                 <NA>
#> 90               Reactive devaluation                 <NA>
#> 91                   Recency illusion                 <NA>
#> 92                     Restraint bias                 <NA>
#> 93             Rhyme as reason effect          mw-redirect
#> 94                  Risk compensation                 <NA>
#> 95                      Salience bias          mw-redirect
#> 96                     Selection bias                 <NA>
#> 97               Selective perception                 <NA>
#> 98                  Semmelweis reflex                 <NA>
#> 99         Sexual overperception bias          mw-redirect
#> 100            Social comparison bias                 <NA>
#> 101          Social desirability bias                 <NA>
#> 102                   Status quo bias                 <NA>
#> 103                      Stereotyping          mw-redirect
#> 104              Subadditivity effect                 <NA>
#> 105             Subjective validation                 <NA>
#> 106                       Surrogation                 <NA>
#> 107                 Survivorship bias                 <NA>
#> 108                  Time-saving bias                 <NA>
#> 109               Third-person effect                 <NA>
#> 110     Parkinson's law of triviality          mw-redirect
#> 111                 Weber–Fechner law                 <NA>
#> 112        Well travelled road effect                 <NA>
#> 113        Women are wonderful effect                 <NA>
#> 114                    Zero-risk bias                 <NA>
#> 115                 Zero-sum thinking                 <NA>
#> 116               Actor-observer bias          mw-redirect
#> 117                    Authority bias                 <NA>
#> 118                Cheerleader effect                 <NA>
#> 119  Defensive attribution hypothesis                 <NA>
#> 120                   Egocentric bias                 <NA>
#> 121         Extrinsic incentives bias                 <NA>
#> 122            False consensus effect                 <NA>
#> 123             False uniqueness bias          mw-redirect
#> 124     Fundamental attribution error                 <NA>
#> 125           Group attribution error                 <NA>
#> 126                       Halo effect                 <NA>
#> 127    Illusion of asymmetric insight                 <NA>
#> 128       Illusion of external agency                 <NA>
#> 129          Illusion of transparency                 <NA>
#> 130              Illusory superiority                 <NA>
#> 131                      Ingroup bias          mw-redirect
#> 132             Just-world hypothesis                 <NA>
#> 133                        Moral luck                 <NA>
#> 134                    Naïve cynicism                 <NA>
#> 135        Naïve realism (psychology)                 <NA>
#> 136         Outgroup homogeneity bias          mw-redirect
#> 137                 Self-serving bias                 <NA>
#> 138           Shared information bias                 <NA>
#> 139              System justification                 <NA>
#> 140             Trait ascription bias                 <NA>
#> 141        Ultimate attribution error                 <NA>
#> 142         Worse-than-average effect                 <NA>
#> 143                Bizarreness effect                 <NA>
#> 144          Cue-dependent forgetting                 <NA>
#> 145                 Cross-race effect                 <NA>
#> 146                      Cryptomnesia                 <NA>
#> 147                   Egocentric bias                 <NA>
#> 148                Fading affect bias                 <NA>
#> 149                      False memory                 <NA>
#> 150                 Generation effect                 <NA>
#> 151                     Google effect                 <NA>
#> 152           Leveling and sharpening                 <NA>
#> 153       Levels-of-processing effect                 <NA>
#> 154             Misinformation effect                 <NA>
#> 155                   Modality effect                 <NA>
#> 156          Cue-dependent forgetting                 <NA>
#> 157           Part-list cueing effect          mw-redirect
#> 158                     Peak-end rule          mw-redirect
#> 159        Picture superiority effect                 <NA>
#> 160 Socioemotional selectivity theory                 <NA>
#> 161                    Primacy effect          mw-redirect
#> 162                    Recency effect          mw-redirect
#> 163            Serial position effect          mw-redirect
#> 164                 Reminiscence bump                 <NA>
#> 165          Misattribution of memory                 <NA>
#> 166                    Spacing effect                 <NA>
#> 167                  Spotlight effect                 <NA>
#> 168                    Suggestibility                 <NA>
#> 169                      Tachypsychia                 <NA>
#> 170                Telescoping effect                 <NA>
#> 171                    Testing effect                 <NA>
#> 172                 Tip of the tongue                 <NA>
#> 173               Von Restorff effect                 <NA>
#> 174                  Zeigarnik effect                 <NA>
#> 175                              <NA> mw-selflink selflink
#> 176             List of memory biases                 <NA>
#>                                           .text
#> 1                              Ambiguity effect
#> 2                                     Anchoring
#> 3                      Anthropocentric thinking
#> 4                              Anthropomorphism
#> 5                              Attentional bias
#> 6                               Automation bias
#> 7                        Availability heuristic
#> 8                          Availability cascade
#> 9                               Backfire effect
#> 10                             Bandwagon effect
#> 11                            Base rate fallacy
#> 12                                  Belief bias
#> 13                          Ben Franklin effect
#> 14                            Berkson's paradox
#> 15                              Bias blind spot
#> 16                       Choice-supportive bias
#> 17                          Clustering illusion
#> 18                            Confirmation bias
#> 19                              Congruence bias
#> 20                          Conjunction fallacy
#> 21               Conservatism (belief revision)
#> 22                   Continued influence effect
#> 23                              Contrast effect
#> 24                                Courtesy bias
#> 25                           Curse of knowledge
#> 26                                    Declinism
#> 27                                 Decoy effect
#> 28                               Default effect
#> 29                          Denomination effect
#> 30                           Disposition effect
#> 31                             Distinction bias
#> 32                        Dunning–Kruger effect
#> 33                             Duration neglect
#> 34                                  Empathy gap
#> 35                             Endowment effect
#> 36                               Experimenter's
#> 37                             expectation bias
#> 38                              Focusing effect
#> 39                                 Forer effect
#> 40                                Barnum effect
#> 41                               Framing effect
#> 42  Frequency illusion or Baader–Meinhof effect
#> 43                         Functional fixedness
#> 44                            Gambler's fallacy
#> 45                                   Groupthink
#> 46                             Hard–easy effect
#> 47                               Hindsight bias
#> 48                     Hostile attribution bias
#> 49                             Hot-hand fallacy
#> 50                       Hyperbolic discounting
#> 51                   Identifiable victim effect
#> 52                                  IKEA effect
#> 53                         Illicit transference
#> 54                          Illusion of control
#> 55                         Illusion of validity
#> 56                         Illusory correlation
#> 57                        Illusory truth effect
#> 58                                  Impact bias
#> 59                         Implicit association
#> 60                             Information bias
#> 61                 Insensitivity to sample size
#> 62                        Irrational escalation
#> 63                        Law of the instrument
#> 64                        Less-is-better effect
#> 65                        Look-elsewhere effect
#> 66                                Loss aversion
#> 67                         Mere exposure effect
#> 68                               Money illusion
#> 69                      Moral credential effect
#> 70                              Negativity bias
#> 71                       Neglect of probability
#> 72                                Normalcy bias
#> 73                            Not invented here
#> 74                   Observer-expectancy effect
#> 75                                Omission bias
#> 76                                Optimism bias
#> 77                               Ostrich effect
#> 78                                 Outcome bias
#> 79                        Overconfidence effect
#> 80                                   Pareidolia
#> 81                             Pygmalion effect
#> 82                               Pessimism bias
#> 83                             Planning fallacy
#> 84                Post-purchase rationalization
#> 85                                 Present bias
#> 86                          Pro-innovation bias
#> 87                              Projection bias
#> 88                       Pseudocertainty effect
#> 89                                    Reactance
#> 90                         Reactive devaluation
#> 91                             Recency illusion
#> 92                               Restraint bias
#> 93                       Rhyme as reason effect
#> 94                            Risk compensation
#> 95                                Salience bias
#> 96                               Selection bias
#> 97                         Selective perception
#> 98                            Semmelweis reflex
#> 99                   Sexual overperception bias
#> 100                      Social comparison bias
#> 101                    Social desirability bias
#> 102                             Status quo bias
#> 103                                Stereotyping
#> 104                        Subadditivity effect
#> 105                       Subjective validation
#> 106                                 Surrogation
#> 107                           Survivorship bias
#> 108                            Time-saving bias
#> 109                         Third-person effect
#> 110               Parkinson's law of triviality
#> 111                           Weber–Fechner law
#> 112                  Well travelled road effect
#> 113                  Women are wonderful effect
#> 114                              Zero-risk bias
#> 115                               Zero-sum bias
#> 116                         Actor-observer bias
#> 117                              Authority bias
#> 118                          Cheerleader effect
#> 119            Defensive attribution hypothesis
#> 120                             Egocentric bias
#> 121                   Extrinsic incentives bias
#> 122                      False consensus effect
#> 123                       False uniqueness bias
#> 124               Fundamental attribution error
#> 125                     Group attribution error
#> 126                                 Halo effect
#> 127              Illusion of asymmetric insight
#> 128                 Illusion of external agency
#> 129                    Illusion of transparency
#> 130                        Illusory superiority
#> 131                                Ingroup bias
#> 132                       Just-world hypothesis
#> 133                                  Moral luck
#> 134                              Naïve cynicism
#> 135                               Naïve realism
#> 136                   Outgroup homogeneity bias
#> 137                           Self-serving bias
#> 138                     Shared information bias
#> 139                        System justification
#> 140                       Trait ascription bias
#> 141                  Ultimate attribution error
#> 142                   Worse-than-average effect
#> 143                          Bizarreness effect
#> 144                              Context effect
#> 145                           Cross-race effect
#> 146                                Cryptomnesia
#> 147                             Egocentric bias
#> 148                          Fading affect bias
#> 149                                False memory
#> 150                           Generation effect
#> 151                               Google effect
#> 152                     Leveling and sharpening
#> 153                 Levels-of-processing effect
#> 154                       Misinformation effect
#> 155                             Modality effect
#> 156                  Mood-congruent memory bias
#> 157                     Part-list cueing effect
#> 158                               Peak-end rule
#> 159                  Picture superiority effect
#> 160           Socioemotional selectivity theory
#> 161                              Primacy effect
#> 162                              recency effect
#> 163                      serial position effect
#> 164                           Reminiscence bump
#> 165                            Source confusion
#> 166                              Spacing effect
#> 167                            Spotlight effect
#> 168                              Suggestibility
#> 169                                Tachypsychia
#> 170                          Telescoping effect
#> 171                              Testing effect
#> 172                           Tip of the tongue
#> 173                         von Restorff effect
#> 174                            Zeigarnik effect
#> 175                                     General
#> 176                                      Memory