Convert collection of html nodes into data frame

Convert collection of html nodes into data frame

html_attrs_dfr(x, attrs = NULL, trim = FALSE,
  defaults = NA_character_, add_text = TRUE)

Arguments

x
xml_nodeset object, containing text and attributes of interest

attrs
character vector of attribute names. If missing, all attributes will be used

trim
if TRUE, will trim leading and trailing spaces

defaults
character vector of default values to be passed to rvest::html_attr(). Recycled to match length of attrs

add_text
if TRUE, node content will be added as .text column (using rvest::html_text)

Value

data frame with one row per xml node, consisting of an html_text column with text and additional columns with attributes

Examples

library(polite)
library(rvest)

#> Loading required package: xml2

bow("https://en.wikipedia.org/wiki/List_of_cognitive_biases") %>%
  scrape() %>%
  html_nodes("tr td:nth-child(1) a") %>%
  html_attrs_dfr()

#>                                                      href
#> 1                                  /wiki/Ambiguity_effect
#> 2                                         /wiki/Anchoring
#> 3                          /wiki/Anthropocentric_thinking
#> 4   /wiki/Anthropomorphism#Psychology_of_Anthropomorphism
#> 5                                  /wiki/Attentional_bias
#> 6                            /wiki/Attribute_substitution
#> 7                                   /wiki/Automation_bias
#> 8                            /wiki/Availability_heuristic
#> 9                              /wiki/Availability_cascade
#> 10                                  /wiki/Backfire_effect
#> 11                                 /wiki/Bandwagon_effect
#> 12                                /wiki/Base_rate_fallacy
#> 13                                      /wiki/Belief_bias
#> 14                              /wiki/Ben_Franklin_effect
#> 15                              /wiki/Berkson%27s_paradox
#> 16                                  /wiki/Bias_blind_spot
#> 17                           /wiki/Choice-supportive_bias
#> 18                              /wiki/Clustering_illusion
#> 19                                  /wiki/Compassion_fade
#> 20                                /wiki/Confirmation_bias
#> 21                                  /wiki/Congruence_bias
#> 22                              /wiki/Conjunction_fallacy
#> 23                   /wiki/Conservatism_(belief_revision)
#> 24                       /wiki/Continued_influence_effect
#> 25                                  /wiki/Contrast_effect
#> 26                                  /wiki/Shy_Tory_Factor
#> 27                               /wiki/Curse_of_knowledge
#> 28                                        /wiki/Declinism
#> 29                                     /wiki/Decoy_effect
#> 30                                   /wiki/Default_effect
#> 31                              /wiki/Denomination_effect
#> 32                               /wiki/Disposition_effect
#> 33                                 /wiki/Distinction_bias
#> 34                    /wiki/Dunning%E2%80%93Kruger_effect
#> 35                                 /wiki/Duration_neglect
#> 36                                      /wiki/Empathy_gap
#> 37                          /wiki/End-of-history_illusion
#> 38                                 /wiki/Endowment_effect
#> 39                            /wiki/Experimenter%27s_bias
#> 40                                 /wiki/Expectation_bias
#> 41                                  /wiki/Focusing_effect
#> 42                                     /wiki/Forer_effect
#> 43                                    /wiki/Barnum_effect
#> 44                      /wiki/Framing_effect_(psychology)
#> 45                             /wiki/Functional_fixedness
#> 46                              /wiki/Gambler%27s_fallacy
#> 47                                       /wiki/Groupthink
#> 48                         /wiki/Hard%E2%80%93easy_effect
#> 49                                   /wiki/Hindsight_bias
#> 50                         /wiki/Hostile_attribution_bias
#> 51                                 /wiki/Hot-hand_fallacy
#> 52                           /wiki/Hyperbolic_discounting
#> 53                       /wiki/Identifiable_victim_effect
#> 54                                      /wiki/IKEA_effect
#> 55                             /wiki/Illicit_transference
#> 56                              /wiki/Illusion_of_control
#> 57                             /wiki/Illusion_of_validity
#> 58                             /wiki/Illusory_correlation
#> 59                            /wiki/Illusory_truth_effect
#> 60                                      /wiki/Impact_bias
#> 61                        /wiki/Implicit_association_test
#> 62                    /wiki/Information_bias_(psychology)
#> 63                     /wiki/Insensitivity_to_sample_size
#> 64                         /wiki/Escalation_of_commitment
#> 65                            /wiki/Law_of_the_instrument
#> 66                            /wiki/Less-is-better_effect
#> 67                            /wiki/Look-elsewhere_effect
#> 68                                    /wiki/Loss_aversion
#> 69                             /wiki/Mere_exposure_effect
#> 70                                   /wiki/Money_illusion
#> 71                                 /wiki/Moral_credential
#> 72                                  /wiki/Negativity_bias
#> 73                           /wiki/Neglect_of_probability
#> 74                                    /wiki/Normalcy_bias
#> 75                                /wiki/Not_invented_here
#> 76                       /wiki/Observer-expectancy_effect
#> 77                                    /wiki/Omission_bias
#> 78                                    /wiki/Optimism_bias
#> 79                                   /wiki/Ostrich_effect
#> 80                                     /wiki/Outcome_bias
#> 81                            /wiki/Overconfidence_effect
#> 82                                       /wiki/Pareidolia
#> 83                                 /wiki/Pygmalion_effect
#> 84                                   /wiki/Pessimism_bias
#> 85                                 /wiki/Planning_fallacy
#> 86                    /wiki/Post-purchase_rationalization
#> 87                                     /wiki/Present_bias
#> 88                              /wiki/Pro-innovation_bias
#> 89                                  /wiki/Projection_bias
#> 90                           /wiki/Pseudocertainty_effect
#> 91                           /wiki/Reactance_(psychology)
#> 92                             /wiki/Reactive_devaluation
#> 93                                 /wiki/Recency_illusion
#> 94                                   /wiki/Restraint_bias
#> 95                           /wiki/Rhyme_as_reason_effect
#> 96                                /wiki/Risk_compensation
#> 97                                    /wiki/Salience_bias
#> 98                                   /wiki/Selection_bias
#> 99                             /wiki/Selective_perception
#> 100                               /wiki/Semmelweis_reflex
#> 101                      /wiki/Sexual_overperception_bias
#> 102              /wiki/Compassion_fade#Singularity_effect
#> 103                          /wiki/Social_comparison_bias
#> 104                        /wiki/Social_desirability_bias
#> 105                                 /wiki/Status_quo_bias
#> 106                                    /wiki/Stereotyping
#> 107                            /wiki/Subadditivity_effect
#> 108                           /wiki/Subjective_validation
#> 109                                     /wiki/Surrogation
#> 110                               /wiki/Survivorship_bias
#> 111                                /wiki/Time-saving_bias
#> 112                             /wiki/Third-person_effect
#> 113                 /wiki/Parkinson%27s_law_of_triviality
#> 114                       /wiki/Weber%E2%80%93Fechner_law
#> 115                      /wiki/Well_travelled_road_effect
#> 116                      /wiki/Women_are_wonderful_effect
#> 117                                  /wiki/Zero-risk_bias
#> 118                               /wiki/Zero-sum_thinking
#> 119                             /wiki/Actor-observer_bias
#> 120                                  /wiki/Authority_bias
#> 121                              /wiki/Cheerleader_effect
#> 122                /wiki/Defensive_attribution_hypothesis
#> 123                                 /wiki/Egocentric_bias
#> 124                       /wiki/Extrinsic_incentives_bias
#> 125                          /wiki/False_consensus_effect
#> 126                           /wiki/False_uniqueness_bias
#> 127                   /wiki/Fundamental_attribution_error
#> 128                         /wiki/Group_attribution_error
#> 129                                     /wiki/Halo_effect
#> 130                  /wiki/Illusion_of_asymmetric_insight
#> 131                     /wiki/Illusion_of_external_agency
#> 132                        /wiki/Illusion_of_transparency
#> 133                            /wiki/Illusory_superiority
#> 134                                    /wiki/Ingroup_bias
#> 135                           /wiki/Just-world_hypothesis
#> 136                                      /wiki/Moral_luck
#> 137                             /wiki/Na%C3%AFve_cynicism
#> 138                 /wiki/Na%C3%AFve_realism_(psychology)
#> 139                       /wiki/Outgroup_homogeneity_bias
#> 140                               /wiki/Self-serving_bias
#> 141                         /wiki/Shared_information_bias
#> 142                            /wiki/System_justification
#> 143                           /wiki/Trait_ascription_bias
#> 144                      /wiki/Ultimate_attribution_error
#> 145                       /wiki/Worse-than-average_effect
#> 146                              /wiki/Bizarreness_effect
#> 147                        /wiki/Cue-dependent_forgetting
#> 148                               /wiki/Cross-race_effect
#> 149                                    /wiki/Cryptomnesia
#> 150                                 /wiki/Egocentric_bias
#> 151                              /wiki/Fading_affect_bias
#> 152                                    /wiki/False_memory
#> 153                               /wiki/Generation_effect
#> 154                                   /wiki/Google_effect
#> 155                         /wiki/Leveling_and_sharpening
#> 156                     /wiki/Levels-of-processing_effect
#> 157                           /wiki/Misinformation_effect
#> 158                                 /wiki/Modality_effect
#> 159                        /wiki/Cue-dependent_forgetting
#> 160                         /wiki/Part-list_cueing_effect
#> 161                                   /wiki/Peak-end_rule
#> 162                      /wiki/Picture_superiority_effect
#> 163               /wiki/Socioemotional_selectivity_theory
#> 164                                  /wiki/Primacy_effect
#> 165                                  /wiki/Recency_effect
#> 166                          /wiki/Serial_position_effect
#> 167                               /wiki/Reminiscence_bump
#> 168                        /wiki/Misattribution_of_memory
#> 169                                  /wiki/Spacing_effect
#> 170                                /wiki/Spotlight_effect
#> 171                         /wiki/Suggestibility#External
#> 172                                    /wiki/Tachypsychia
#> 173                              /wiki/Telescoping_effect
#> 174                                  /wiki/Testing_effect
#> 175                               /wiki/Tip_of_the_tongue
#> 176                             /wiki/Von_Restorff_effect
#> 177                                /wiki/Zeigarnik_effect
#> 178                                                  <NA>
#> 179                           /wiki/List_of_memory_biases
#>                                 title                class
#> 1                    Ambiguity effect                 <NA>
#> 2                           Anchoring                 <NA>
#> 3            Anthropocentric thinking          mw-redirect
#> 4                    Anthropomorphism                 <NA>
#> 5                    Attentional bias                 <NA>
#> 6              Attribute substitution                 <NA>
#> 7                     Automation bias                 <NA>
#> 8              Availability heuristic                 <NA>
#> 9                Availability cascade                 <NA>
#> 10                    Backfire effect          mw-redirect
#> 11                   Bandwagon effect                 <NA>
#> 12                  Base rate fallacy                 <NA>
#> 13                        Belief bias                 <NA>
#> 14                Ben Franklin effect                 <NA>
#> 15                  Berkson's paradox                 <NA>
#> 16                    Bias blind spot                 <NA>
#> 17             Choice-supportive bias                 <NA>
#> 18                Clustering illusion                 <NA>
#> 19                    Compassion fade                 <NA>
#> 20                  Confirmation bias                 <NA>
#> 21                    Congruence bias                 <NA>
#> 22                Conjunction fallacy                 <NA>
#> 23     Conservatism (belief revision)                 <NA>
#> 24         Continued influence effect          mw-redirect
#> 25                    Contrast effect                 <NA>
#> 26                    Shy Tory Factor          mw-redirect
#> 27                 Curse of knowledge                 <NA>
#> 28                          Declinism                 <NA>
#> 29                       Decoy effect                 <NA>
#> 30                     Default effect                 <NA>
#> 31                Denomination effect                 <NA>
#> 32                 Disposition effect                 <NA>
#> 33                   Distinction bias                 <NA>
#> 34              Dunning–Kruger effect                 <NA>
#> 35                   Duration neglect                 <NA>
#> 36                        Empathy gap                 <NA>
#> 37            End-of-history illusion                 <NA>
#> 38                   Endowment effect                 <NA>
#> 39                Experimenter's bias          mw-redirect
#> 40                   Expectation bias          mw-redirect
#> 41                    Focusing effect          mw-redirect
#> 42                       Forer effect          mw-redirect
#> 43                      Barnum effect                 <NA>
#> 44        Framing effect (psychology)                 <NA>
#> 45               Functional fixedness                 <NA>
#> 46                  Gambler's fallacy                 <NA>
#> 47                         Groupthink                 <NA>
#> 48                   Hard–easy effect                 <NA>
#> 49                     Hindsight bias                 <NA>
#> 50           Hostile attribution bias                 <NA>
#> 51                   Hot-hand fallacy          mw-redirect
#> 52             Hyperbolic discounting                 <NA>
#> 53         Identifiable victim effect                 <NA>
#> 54                        IKEA effect                 <NA>
#> 55               Illicit transference          mw-redirect
#> 56                Illusion of control                 <NA>
#> 57               Illusion of validity                 <NA>
#> 58               Illusory correlation                 <NA>
#> 59              Illusory truth effect                 <NA>
#> 60                        Impact bias                 <NA>
#> 61          Implicit association test          mw-redirect
#> 62      Information bias (psychology)                 <NA>
#> 63       Insensitivity to sample size                 <NA>
#> 64           Escalation of commitment                 <NA>
#> 65              Law of the instrument                 <NA>
#> 66              Less-is-better effect                 <NA>
#> 67              Look-elsewhere effect                 <NA>
#> 68                      Loss aversion                 <NA>
#> 69               Mere exposure effect          mw-redirect
#> 70                     Money illusion                 <NA>
#> 71                   Moral credential          mw-redirect
#> 72                    Negativity bias                 <NA>
#> 73             Neglect of probability                 <NA>
#> 74                      Normalcy bias                 <NA>
#> 75                  Not invented here                 <NA>
#> 76         Observer-expectancy effect                 <NA>
#> 77                      Omission bias                 <NA>
#> 78                      Optimism bias                 <NA>
#> 79                     Ostrich effect                 <NA>
#> 80                       Outcome bias                 <NA>
#> 81              Overconfidence effect                 <NA>
#> 82                         Pareidolia                 <NA>
#> 83                   Pygmalion effect                 <NA>
#> 84                     Pessimism bias          mw-redirect
#> 85                   Planning fallacy                 <NA>
#> 86      Post-purchase rationalization          mw-redirect
#> 87                       Present bias                 <NA>
#> 88                Pro-innovation bias                 <NA>
#> 89                    Projection bias          mw-redirect
#> 90             Pseudocertainty effect                 <NA>
#> 91             Reactance (psychology)                 <NA>
#> 92               Reactive devaluation                 <NA>
#> 93                   Recency illusion                 <NA>
#> 94                     Restraint bias                 <NA>
#> 95             Rhyme as reason effect          mw-redirect
#> 96                  Risk compensation                 <NA>
#> 97                      Salience bias          mw-redirect
#> 98                     Selection bias                 <NA>
#> 99               Selective perception                 <NA>
#> 100                 Semmelweis reflex                 <NA>
#> 101        Sexual overperception bias          mw-redirect
#> 102                   Compassion fade                 <NA>
#> 103            Social comparison bias                 <NA>
#> 104          Social desirability bias                 <NA>
#> 105                   Status quo bias                 <NA>
#> 106                      Stereotyping          mw-redirect
#> 107              Subadditivity effect                 <NA>
#> 108             Subjective validation                 <NA>
#> 109                       Surrogation                 <NA>
#> 110                 Survivorship bias                 <NA>
#> 111                  Time-saving bias                 <NA>
#> 112               Third-person effect                 <NA>
#> 113     Parkinson's law of triviality          mw-redirect
#> 114                 Weber–Fechner law                 <NA>
#> 115        Well travelled road effect                 <NA>
#> 116        Women are wonderful effect                 <NA>
#> 117                    Zero-risk bias                 <NA>
#> 118                 Zero-sum thinking                 <NA>
#> 119               Actor-observer bias          mw-redirect
#> 120                    Authority bias                 <NA>
#> 121                Cheerleader effect                 <NA>
#> 122  Defensive attribution hypothesis                 <NA>
#> 123                   Egocentric bias                 <NA>
#> 124         Extrinsic incentives bias                 <NA>
#> 125            False consensus effect                 <NA>
#> 126             False uniqueness bias          mw-redirect
#> 127     Fundamental attribution error                 <NA>
#> 128           Group attribution error                 <NA>
#> 129                       Halo effect                 <NA>
#> 130    Illusion of asymmetric insight                 <NA>
#> 131       Illusion of external agency                 <NA>
#> 132          Illusion of transparency                 <NA>
#> 133              Illusory superiority                 <NA>
#> 134                      Ingroup bias          mw-redirect
#> 135             Just-world hypothesis                 <NA>
#> 136                        Moral luck                 <NA>
#> 137                    Naïve cynicism                 <NA>
#> 138        Naïve realism (psychology)                 <NA>
#> 139         Outgroup homogeneity bias          mw-redirect
#> 140                 Self-serving bias                 <NA>
#> 141           Shared information bias                 <NA>
#> 142              System justification                 <NA>
#> 143             Trait ascription bias                 <NA>
#> 144        Ultimate attribution error                 <NA>
#> 145         Worse-than-average effect                 <NA>
#> 146                Bizarreness effect                 <NA>
#> 147          Cue-dependent forgetting                 <NA>
#> 148                 Cross-race effect                 <NA>
#> 149                      Cryptomnesia                 <NA>
#> 150                   Egocentric bias                 <NA>
#> 151                Fading affect bias                 <NA>
#> 152                      False memory                 <NA>
#> 153                 Generation effect                 <NA>
#> 154                     Google effect                 <NA>
#> 155           Leveling and sharpening                 <NA>
#> 156       Levels-of-processing effect                 <NA>
#> 157             Misinformation effect                 <NA>
#> 158                   Modality effect                 <NA>
#> 159          Cue-dependent forgetting                 <NA>
#> 160           Part-list cueing effect          mw-redirect
#> 161                     Peak-end rule          mw-redirect
#> 162        Picture superiority effect                 <NA>
#> 163 Socioemotional selectivity theory                 <NA>
#> 164                    Primacy effect          mw-redirect
#> 165                    Recency effect          mw-redirect
#> 166            Serial position effect          mw-redirect
#> 167                 Reminiscence bump                 <NA>
#> 168          Misattribution of memory                 <NA>
#> 169                    Spacing effect                 <NA>
#> 170                  Spotlight effect                 <NA>
#> 171                    Suggestibility                 <NA>
#> 172                      Tachypsychia                 <NA>
#> 173                Telescoping effect                 <NA>
#> 174                    Testing effect                 <NA>
#> 175                 Tip of the tongue                 <NA>
#> 176               Von Restorff effect                 <NA>
#> 177                  Zeigarnik effect                 <NA>
#> 178                              <NA> mw-selflink selflink
#> 179             List of memory biases                 <NA>
#>                                                 .text
#> 1                                    Ambiguity effect
#> 2                                           Anchoring
#> 3                            Anthropocentric thinking
#> 4                                    Anthropomorphism
#> 5                                    Attentional bias
#> 6                              Attribute substitution
#> 7                                     Automation bias
#> 8                              Availability heuristic
#> 9                                Availability cascade
#> 10                                    Backfire effect
#> 11                                   Bandwagon effect
#> 12                                  Base rate fallacy
#> 13                                        Belief bias
#> 14                                Ben Franklin effect
#> 15                                  Berkson's paradox
#> 16                                    Bias blind spot
#> 17                             Choice-supportive bias
#> 18                                Clustering illusion
#> 19                                    Compassion fade
#> 20                                  Confirmation bias
#> 21                                    Congruence bias
#> 22                                Conjunction fallacy
#> 23                     Conservatism (belief revision)
#> 24                         Continued influence effect
#> 25                                    Contrast effect
#> 26                                      Courtesy bias
#> 27                                 Curse of knowledge
#> 28                                          Declinism
#> 29                                       Decoy effect
#> 30                                     Default effect
#> 31                                Denomination effect
#> 32                                 Disposition effect
#> 33                                   Distinction bias
#> 34                              Dunning–Kruger effect
#> 35                                   Duration neglect
#> 36                                        Empathy gap
#> 37                            End-of-history illusion
#> 38                                   Endowment effect
#> 39                                     Experimenter's
#> 40                                   expectation bias
#> 41                                    Focusing effect
#> 42                                       Forer effect
#> 43                                      Barnum effect
#> 44                                     Framing effect
#> 45                               Functional fixedness
#> 46                                  Gambler's fallacy
#> 47                                         Groupthink
#> 48                                   Hard–easy effect
#> 49                                     Hindsight bias
#> 50                           Hostile attribution bias
#> 51                                   Hot-hand fallacy
#> 52                             Hyperbolic discounting
#> 53                         Identifiable victim effect
#> 54                                        IKEA effect
#> 55                               Illicit transference
#> 56                                Illusion of control
#> 57                               Illusion of validity
#> 58                               Illusory correlation
#> 59                              Illusory truth effect
#> 60                                        Impact bias
#> 61                               Implicit association
#> 62                                   Information bias
#> 63                       Insensitivity to sample size
#> 64  Irrational escalation or Escalation of commitment
#> 65                              Law of the instrument
#> 66                              Less-is-better effect
#> 67                              Look-elsewhere effect
#> 68                                      Loss aversion
#> 69                               Mere exposure effect
#> 70                                     Money illusion
#> 71                            Moral credential effect
#> 72                                    Negativity bias
#> 73                             Neglect of probability
#> 74                                      Normalcy bias
#> 75                                  Not invented here
#> 76                         Observer-expectancy effect
#> 77                                      Omission bias
#> 78                                      Optimism bias
#> 79                                     Ostrich effect
#> 80                                       Outcome bias
#> 81                              Overconfidence effect
#> 82                                         Pareidolia
#> 83                                   Pygmalion effect
#> 84                                     Pessimism bias
#> 85                                   Planning fallacy
#> 86                      Post-purchase rationalization
#> 87                                       Present bias
#> 88                                Pro-innovation bias
#> 89                                    Projection bias
#> 90                             Pseudocertainty effect
#> 91                                          Reactance
#> 92                               Reactive devaluation
#> 93                                   Recency illusion
#> 94                                     Restraint bias
#> 95                             Rhyme as reason effect
#> 96                                  Risk compensation
#> 97                                      Salience bias
#> 98                                     Selection bias
#> 99                               Selective perception
#> 100                                 Semmelweis reflex
#> 101                        Sexual overperception bias
#> 102                                Singularity effect
#> 103                            Social comparison bias
#> 104                          Social desirability bias
#> 105                                   Status quo bias
#> 106                                      Stereotyping
#> 107                              Subadditivity effect
#> 108                             Subjective validation
#> 109                                       Surrogation
#> 110                                 Survivorship bias
#> 111                                  Time-saving bias
#> 112                               Third-person effect
#> 113                     Parkinson's law of triviality
#> 114                                 Weber–Fechner law
#> 115                        Well travelled road effect
#> 116                        Women are wonderful effect
#> 117                                    Zero-risk bias
#> 118                                     Zero-sum bias
#> 119                               Actor-observer bias
#> 120                                    Authority bias
#> 121                                Cheerleader effect
#> 122                  Defensive attribution hypothesis
#> 123                                   Egocentric bias
#> 124                         Extrinsic incentives bias
#> 125                            False consensus effect
#> 126                             False uniqueness bias
#> 127                     Fundamental attribution error
#> 128                           Group attribution error
#> 129                                       Halo effect
#> 130                    Illusion of asymmetric insight
#> 131                       Illusion of external agency
#> 132                          Illusion of transparency
#> 133                              Illusory superiority
#> 134                                      Ingroup bias
#> 135                             Just-world hypothesis
#> 136                                        Moral luck
#> 137                                    Naïve cynicism
#> 138                                     Naïve realism
#> 139                         Outgroup homogeneity bias
#> 140                                 Self-serving bias
#> 141                           Shared information bias
#> 142                              System justification
#> 143                             Trait ascription bias
#> 144                        Ultimate attribution error
#> 145                         Worse-than-average effect
#> 146                                Bizarreness effect
#> 147                                    Context effect
#> 148                                 Cross-race effect
#> 149                                      Cryptomnesia
#> 150                                   Egocentric bias
#> 151                                Fading affect bias
#> 152                                      False memory
#> 153                                 Generation effect
#> 154                                     Google effect
#> 155                           Leveling and sharpening
#> 156                       Levels-of-processing effect
#> 157                             Misinformation effect
#> 158                                   Modality effect
#> 159                        Mood-congruent memory bias
#> 160                           Part-list cueing effect
#> 161                                     Peak-end rule
#> 162                        Picture superiority effect
#> 163                 Socioemotional selectivity theory
#> 164                                    Primacy effect
#> 165                                    recency effect
#> 166                            serial position effect
#> 167                                 Reminiscence bump
#> 168                                  Source confusion
#> 169                                    Spacing effect
#> 170                                  Spotlight effect
#> 171                                    Suggestibility
#> 172                                      Tachypsychia
#> 173                                Telescoping effect
#> 174                                    Testing effect
#> 175                                 Tip of the tongue
#> 176                               von Restorff effect
#> 177                                  Zeigarnik effect
#> 178                                           General
#> 179                                            Memory