Convert collection of html nodes into data frame

Convert collection of html nodes into data frame

html_attrs_dfr(x, attrs = NULL, trim = FALSE,
  defaults = NA_character_, add_text = TRUE)

Arguments

x
xml_nodeset object, containing text and attributes of interest

attrs
character vector of attribute names. If missing, all attributes will be used

trim
if TRUE, will trim leading and trailing spaces

defaults
character vector of default values to be passed to rvest::html_attr(). Recycled to match length of attrs

add_text
if TRUE, node content will be added as .text column (using rvest::html_text)

Value

data frame with one row per xml node, consisting of an html_text column with text and additional columns with attributes

Examples

# \donttest{
library(polite)
library(rvest)

#> Loading required package: xml2

bow("https://en.wikipedia.org/wiki/List_of_cognitive_biases") %>%
  scrape() %>%
  html_nodes("tr td:nth-child(1) a") %>%
  html_attrs_dfr()

#>                                                      href
#> 1                                  /wiki/Ambiguity_effect
#> 2                                         /wiki/Anchoring
#> 3                          /wiki/Anthropocentric_thinking
#> 4   /wiki/Anthropomorphism#Psychology_of_Anthropomorphism
#> 5                                  /wiki/Attentional_bias
#> 6                            /wiki/Attribute_substitution
#> 7                                   /wiki/Automation_bias
#> 8                            /wiki/Availability_heuristic
#> 9                              /wiki/Availability_cascade
#> 10                                  /wiki/Backfire_effect
#> 11                                 /wiki/Bandwagon_effect
#> 12                                /wiki/Base_rate_fallacy
#> 13                                      /wiki/Belief_bias
#> 14                              /wiki/Ben_Franklin_effect
#> 15                              /wiki/Berkson%27s_paradox
#> 16                                  /wiki/Bias_blind_spot
#> 17                           /wiki/Choice-supportive_bias
#> 18                              /wiki/Clustering_illusion
#> 19                                  /wiki/Compassion_fade
#> 20                                /wiki/Confirmation_bias
#> 21                                  /wiki/Congruence_bias
#> 22                              /wiki/Conjunction_fallacy
#> 23                   /wiki/Conservatism_(belief_revision)
#> 24                       /wiki/Continued_influence_effect
#> 25                                  /wiki/Contrast_effect
#> 26                               /wiki/Curse_of_knowledge
#> 27                                        /wiki/Declinism
#> 28                                     /wiki/Decoy_effect
#> 29                                   /wiki/Default_effect
#> 30                              /wiki/Denomination_effect
#> 31                               /wiki/Disposition_effect
#> 32                                 /wiki/Distinction_bias
#> 33                    /wiki/Dunning%E2%80%93Kruger_effect
#> 34                                 /wiki/Duration_neglect
#> 35                                      /wiki/Empathy_gap
#> 36                          /wiki/End-of-history_illusion
#> 37                                 /wiki/Endowment_effect
#> 38                            /wiki/Experimenter%27s_bias
#> 39                                 /wiki/Expectation_bias
#> 40                                  /wiki/Focusing_effect
#> 41                                     /wiki/Forer_effect
#> 42                                    /wiki/Barnum_effect
#> 43                      /wiki/Framing_effect_(psychology)
#> 44                             /wiki/Functional_fixedness
#> 45                              /wiki/Gambler%27s_fallacy
#> 46                                       /wiki/Groupthink
#> 47                         /wiki/Hard%E2%80%93easy_effect
#> 48                                   /wiki/Hindsight_bias
#> 49                         /wiki/Hostile_attribution_bias
#> 50                                 /wiki/Hot-hand_fallacy
#> 51                           /wiki/Hyperbolic_discounting
#> 52                       /wiki/Identifiable_victim_effect
#> 53                                      /wiki/IKEA_effect
#> 54                             /wiki/Illicit_transference
#> 55                              /wiki/Illusion_of_control
#> 56                             /wiki/Illusion_of_validity
#> 57                             /wiki/Illusory_correlation
#> 58                            /wiki/Illusory_truth_effect
#> 59                                      /wiki/Impact_bias
#> 60                        /wiki/Implicit_association_test
#> 61                    /wiki/Information_bias_(psychology)
#> 62                     /wiki/Insensitivity_to_sample_size
#> 63                         /wiki/Escalation_of_commitment
#> 64                            /wiki/Law_of_the_instrument
#> 65                            /wiki/Less-is-better_effect
#> 66                            /wiki/Look-elsewhere_effect
#> 67                                    /wiki/Loss_aversion
#> 68                             /wiki/Mere_exposure_effect
#> 69                                   /wiki/Money_illusion
#> 70                                 /wiki/Moral_credential
#> 71                                  /wiki/Negativity_bias
#> 72                           /wiki/Neglect_of_probability
#> 73                                    /wiki/Normalcy_bias
#> 74                                /wiki/Not_invented_here
#> 75                       /wiki/Observer-expectancy_effect
#> 76                                    /wiki/Omission_bias
#> 77                                    /wiki/Optimism_bias
#> 78                                   /wiki/Ostrich_effect
#> 79                                     /wiki/Outcome_bias
#> 80                            /wiki/Overconfidence_effect
#> 81                                       /wiki/Pareidolia
#> 82                                 /wiki/Pygmalion_effect
#> 83                                   /wiki/Pessimism_bias
#> 84                                 /wiki/Planning_fallacy
#> 85                                     /wiki/Present_bias
#> 86                              /wiki/Pro-innovation_bias
#> 87                                  /wiki/Projection_bias
#> 88                           /wiki/Pseudocertainty_effect
#> 89                           /wiki/Reactance_(psychology)
#> 90                             /wiki/Reactive_devaluation
#> 91                                 /wiki/Recency_illusion
#> 92                                   /wiki/Restraint_bias
#> 93                           /wiki/Rhyme_as_reason_effect
#> 94                                /wiki/Risk_compensation
#> 95                                    /wiki/Salience_bias
#> 96                                   /wiki/Selection_bias
#> 97                             /wiki/Selective_perception
#> 98                                /wiki/Semmelweis_reflex
#> 99                       /wiki/Sexual_overperception_bias
#> 100              /wiki/Compassion_fade#Singularity_effect
#> 101                          /wiki/Social_comparison_bias
#> 102                        /wiki/Social_desirability_bias
#> 103                                 /wiki/Status_quo_bias
#> 104                                    /wiki/Stereotyping
#> 105                            /wiki/Subadditivity_effect
#> 106                           /wiki/Subjective_validation
#> 107                                     /wiki/Surrogation
#> 108                               /wiki/Survivorship_bias
#> 109                                /wiki/Time-saving_bias
#> 110                             /wiki/Third-person_effect
#> 111                 /wiki/Parkinson%27s_law_of_triviality
#> 112                       /wiki/Weber%E2%80%93Fechner_law
#> 113                      /wiki/Well_travelled_road_effect
#> 114                      /wiki/Women_are_wonderful_effect
#> 115                                  /wiki/Zero-risk_bias
#> 116                               /wiki/Zero-sum_thinking
#> 117                             /wiki/Actor-observer_bias
#> 118                                  /wiki/Authority_bias
#> 119                              /wiki/Cheerleader_effect
#> 120                /wiki/Defensive_attribution_hypothesis
#> 121                                 /wiki/Egocentric_bias
#> 122                       /wiki/Extrinsic_incentives_bias
#> 123                          /wiki/False_consensus_effect
#> 124                           /wiki/False_uniqueness_bias
#> 125                   /wiki/Fundamental_attribution_error
#> 126                         /wiki/Group_attribution_error
#> 127                                     /wiki/Halo_effect
#> 128                  /wiki/Illusion_of_asymmetric_insight
#> 129                     /wiki/Illusion_of_external_agency
#> 130                        /wiki/Illusion_of_transparency
#> 131                            /wiki/Illusory_superiority
#> 132                                    /wiki/Ingroup_bias
#> 133                           /wiki/Just-world_hypothesis
#> 134                                      /wiki/Moral_luck
#> 135                             /wiki/Na%C3%AFve_cynicism
#> 136                 /wiki/Na%C3%AFve_realism_(psychology)
#> 137                       /wiki/Outgroup_homogeneity_bias
#> 138                               /wiki/Self-serving_bias
#> 139                         /wiki/Shared_information_bias
#> 140                            /wiki/System_justification
#> 141                           /wiki/Trait_ascription_bias
#> 142                      /wiki/Ultimate_attribution_error
#> 143                       /wiki/Worse-than-average_effect
#> 144                              /wiki/Bizarreness_effect
#> 145                        /wiki/Cue-dependent_forgetting
#> 146                               /wiki/Cross-race_effect
#> 147                                    /wiki/Cryptomnesia
#> 148                                 /wiki/Egocentric_bias
#> 149                              /wiki/Fading_affect_bias
#> 150                                    /wiki/False_memory
#> 151                               /wiki/Generation_effect
#> 152                                   /wiki/Google_effect
#> 153                         /wiki/Leveling_and_sharpening
#> 154                     /wiki/Levels-of-processing_effect
#> 155                           /wiki/Misinformation_effect
#> 156                                 /wiki/Modality_effect
#> 157                        /wiki/Cue-dependent_forgetting
#> 158                         /wiki/Part-list_cueing_effect
#> 159                                   /wiki/Peak-end_rule
#> 160                      /wiki/Picture_superiority_effect
#> 161               /wiki/Socioemotional_selectivity_theory
#> 162                                  /wiki/Primacy_effect
#> 163                                  /wiki/Recency_effect
#> 164                          /wiki/Serial_position_effect
#> 165                               /wiki/Reminiscence_bump
#> 166                        /wiki/Misattribution_of_memory
#> 167                                  /wiki/Spacing_effect
#> 168                                /wiki/Spotlight_effect
#> 169                         /wiki/Suggestibility#External
#> 170                                    /wiki/Tachypsychia
#> 171                              /wiki/Telescoping_effect
#> 172                                  /wiki/Testing_effect
#> 173                               /wiki/Tip_of_the_tongue
#> 174                             /wiki/Von_Restorff_effect
#> 175                                /wiki/Zeigarnik_effect
#> 176                                                  <NA>
#> 177                           /wiki/List_of_memory_biases
#>                                 title                class
#> 1                    Ambiguity effect                 <NA>
#> 2                           Anchoring                 <NA>
#> 3            Anthropocentric thinking          mw-redirect
#> 4                    Anthropomorphism                 <NA>
#> 5                    Attentional bias                 <NA>
#> 6              Attribute substitution                 <NA>
#> 7                     Automation bias                 <NA>
#> 8              Availability heuristic                 <NA>
#> 9                Availability cascade                 <NA>
#> 10                    Backfire effect          mw-redirect
#> 11                   Bandwagon effect                 <NA>
#> 12                  Base rate fallacy                 <NA>
#> 13                        Belief bias                 <NA>
#> 14                Ben Franklin effect                 <NA>
#> 15                  Berkson's paradox                 <NA>
#> 16                    Bias blind spot                 <NA>
#> 17             Choice-supportive bias                 <NA>
#> 18                Clustering illusion                 <NA>
#> 19                    Compassion fade                 <NA>
#> 20                  Confirmation bias                 <NA>
#> 21                    Congruence bias                 <NA>
#> 22                Conjunction fallacy                 <NA>
#> 23     Conservatism (belief revision)                 <NA>
#> 24         Continued influence effect          mw-redirect
#> 25                    Contrast effect                 <NA>
#> 26                 Curse of knowledge                 <NA>
#> 27                          Declinism                 <NA>
#> 28                       Decoy effect                 <NA>
#> 29                     Default effect                 <NA>
#> 30                Denomination effect                 <NA>
#> 31                 Disposition effect                 <NA>
#> 32                   Distinction bias                 <NA>
#> 33              Dunning–Kruger effect                 <NA>
#> 34                   Duration neglect                 <NA>
#> 35                        Empathy gap                 <NA>
#> 36            End-of-history illusion                 <NA>
#> 37                   Endowment effect                 <NA>
#> 38                Experimenter's bias          mw-redirect
#> 39                   Expectation bias          mw-redirect
#> 40                    Focusing effect          mw-redirect
#> 41                       Forer effect          mw-redirect
#> 42                      Barnum effect                 <NA>
#> 43        Framing effect (psychology)                 <NA>
#> 44               Functional fixedness                 <NA>
#> 45                  Gambler's fallacy                 <NA>
#> 46                         Groupthink                 <NA>
#> 47                   Hard–easy effect                 <NA>
#> 48                     Hindsight bias                 <NA>
#> 49           Hostile attribution bias                 <NA>
#> 50                   Hot-hand fallacy          mw-redirect
#> 51             Hyperbolic discounting                 <NA>
#> 52         Identifiable victim effect                 <NA>
#> 53                        IKEA effect                 <NA>
#> 54               Illicit transference          mw-redirect
#> 55                Illusion of control                 <NA>
#> 56               Illusion of validity                 <NA>
#> 57               Illusory correlation                 <NA>
#> 58              Illusory truth effect                 <NA>
#> 59                        Impact bias                 <NA>
#> 60          Implicit association test          mw-redirect
#> 61      Information bias (psychology)                 <NA>
#> 62       Insensitivity to sample size                 <NA>
#> 63           Escalation of commitment                 <NA>
#> 64              Law of the instrument                 <NA>
#> 65              Less-is-better effect                 <NA>
#> 66              Look-elsewhere effect                 <NA>
#> 67                      Loss aversion                 <NA>
#> 68               Mere exposure effect          mw-redirect
#> 69                     Money illusion                 <NA>
#> 70                   Moral credential          mw-redirect
#> 71                    Negativity bias                 <NA>
#> 72             Neglect of probability                 <NA>
#> 73                      Normalcy bias                 <NA>
#> 74                  Not invented here                 <NA>
#> 75         Observer-expectancy effect                 <NA>
#> 76                      Omission bias                 <NA>
#> 77                      Optimism bias                 <NA>
#> 78                     Ostrich effect                 <NA>
#> 79                       Outcome bias                 <NA>
#> 80              Overconfidence effect                 <NA>
#> 81                         Pareidolia                 <NA>
#> 82                   Pygmalion effect                 <NA>
#> 83                     Pessimism bias          mw-redirect
#> 84                   Planning fallacy                 <NA>
#> 85                       Present bias                 <NA>
#> 86                Pro-innovation bias                 <NA>
#> 87                    Projection bias          mw-redirect
#> 88             Pseudocertainty effect                 <NA>
#> 89             Reactance (psychology)                 <NA>
#> 90               Reactive devaluation                 <NA>
#> 91                   Recency illusion                 <NA>
#> 92                     Restraint bias                 <NA>
#> 93             Rhyme as reason effect          mw-redirect
#> 94                  Risk compensation                 <NA>
#> 95                      Salience bias          mw-redirect
#> 96                     Selection bias                 <NA>
#> 97               Selective perception                 <NA>
#> 98                  Semmelweis reflex                 <NA>
#> 99         Sexual overperception bias          mw-redirect
#> 100                   Compassion fade                 <NA>
#> 101            Social comparison bias                 <NA>
#> 102          Social desirability bias                 <NA>
#> 103                   Status quo bias                 <NA>
#> 104                      Stereotyping          mw-redirect
#> 105              Subadditivity effect                 <NA>
#> 106             Subjective validation                 <NA>
#> 107                       Surrogation                 <NA>
#> 108                 Survivorship bias                 <NA>
#> 109                  Time-saving bias                 <NA>
#> 110               Third-person effect                 <NA>
#> 111     Parkinson's law of triviality          mw-redirect
#> 112                 Weber–Fechner law                 <NA>
#> 113        Well travelled road effect                 <NA>
#> 114        Women are wonderful effect                 <NA>
#> 115                    Zero-risk bias                 <NA>
#> 116                 Zero-sum thinking                 <NA>
#> 117               Actor-observer bias          mw-redirect
#> 118                    Authority bias                 <NA>
#> 119                Cheerleader effect                 <NA>
#> 120  Defensive attribution hypothesis                 <NA>
#> 121                   Egocentric bias                 <NA>
#> 122         Extrinsic incentives bias                 <NA>
#> 123            False consensus effect                 <NA>
#> 124             False uniqueness bias          mw-redirect
#> 125     Fundamental attribution error                 <NA>
#> 126           Group attribution error                 <NA>
#> 127                       Halo effect                 <NA>
#> 128    Illusion of asymmetric insight                 <NA>
#> 129       Illusion of external agency                 <NA>
#> 130          Illusion of transparency                 <NA>
#> 131              Illusory superiority                 <NA>
#> 132                      Ingroup bias          mw-redirect
#> 133             Just-world hypothesis                 <NA>
#> 134                        Moral luck                 <NA>
#> 135                    Naïve cynicism                 <NA>
#> 136        Naïve realism (psychology)                 <NA>
#> 137         Outgroup homogeneity bias          mw-redirect
#> 138                 Self-serving bias                 <NA>
#> 139           Shared information bias                 <NA>
#> 140              System justification                 <NA>
#> 141             Trait ascription bias                 <NA>
#> 142        Ultimate attribution error                 <NA>
#> 143         Worse-than-average effect                 <NA>
#> 144                Bizarreness effect                 <NA>
#> 145          Cue-dependent forgetting                 <NA>
#> 146                 Cross-race effect                 <NA>
#> 147                      Cryptomnesia                 <NA>
#> 148                   Egocentric bias                 <NA>
#> 149                Fading affect bias                 <NA>
#> 150                      False memory                 <NA>
#> 151                 Generation effect                 <NA>
#> 152                     Google effect                 <NA>
#> 153           Leveling and sharpening                 <NA>
#> 154       Levels-of-processing effect                 <NA>
#> 155             Misinformation effect                 <NA>
#> 156                   Modality effect                 <NA>
#> 157          Cue-dependent forgetting                 <NA>
#> 158           Part-list cueing effect          mw-redirect
#> 159                     Peak-end rule          mw-redirect
#> 160        Picture superiority effect                 <NA>
#> 161 Socioemotional selectivity theory                 <NA>
#> 162                    Primacy effect          mw-redirect
#> 163                    Recency effect          mw-redirect
#> 164            Serial position effect          mw-redirect
#> 165                 Reminiscence bump                 <NA>
#> 166          Misattribution of memory                 <NA>
#> 167                    Spacing effect                 <NA>
#> 168                  Spotlight effect                 <NA>
#> 169                    Suggestibility                 <NA>
#> 170                      Tachypsychia                 <NA>
#> 171                Telescoping effect                 <NA>
#> 172                    Testing effect                 <NA>
#> 173                 Tip of the tongue                 <NA>
#> 174               Von Restorff effect                 <NA>
#> 175                  Zeigarnik effect                 <NA>
#> 176                              <NA> mw-selflink selflink
#> 177             List of memory biases                 <NA>
#>                                                 .text
#> 1                                    Ambiguity effect
#> 2                                           Anchoring
#> 3                            Anthropocentric thinking
#> 4                                    Anthropomorphism
#> 5                                    Attentional bias
#> 6                              Attribute substitution
#> 7                                     Automation bias
#> 8                              Availability heuristic
#> 9                                Availability cascade
#> 10                                    Backfire effect
#> 11                                   Bandwagon effect
#> 12                                  Base rate fallacy
#> 13                                        Belief bias
#> 14                                Ben Franklin effect
#> 15                                  Berkson's paradox
#> 16                                    Bias blind spot
#> 17                             Choice-supportive bias
#> 18                                Clustering illusion
#> 19                                    Compassion fade
#> 20                                  Confirmation bias
#> 21                                    Congruence bias
#> 22                                Conjunction fallacy
#> 23                     Conservatism (belief revision)
#> 24                         Continued influence effect
#> 25                                    Contrast effect
#> 26                                 Curse of knowledge
#> 27                                          Declinism
#> 28                                       Decoy effect
#> 29                                     Default effect
#> 30                                Denomination effect
#> 31                                 Disposition effect
#> 32                                   Distinction bias
#> 33                              Dunning–Kruger effect
#> 34                                   Duration neglect
#> 35                                        Empathy gap
#> 36                            End-of-history illusion
#> 37                                   Endowment effect
#> 38                                     Experimenter's
#> 39                                   expectation bias
#> 40                                    Focusing effect
#> 41                                       Forer effect
#> 42                                      Barnum effect
#> 43                                     Framing effect
#> 44                               Functional fixedness
#> 45                                  Gambler's fallacy
#> 46                                         Groupthink
#> 47                                   Hard–easy effect
#> 48                                     Hindsight bias
#> 49                           Hostile attribution bias
#> 50                                   Hot-hand fallacy
#> 51                             Hyperbolic discounting
#> 52                         Identifiable victim effect
#> 53                                        IKEA effect
#> 54                               Illicit transference
#> 55                                Illusion of control
#> 56                               Illusion of validity
#> 57                               Illusory correlation
#> 58                              Illusory truth effect
#> 59                                        Impact bias
#> 60                               Implicit association
#> 61                                   Information bias
#> 62                       Insensitivity to sample size
#> 63  Irrational escalation or Escalation of commitment
#> 64                              Law of the instrument
#> 65                              Less-is-better effect
#> 66                              Look-elsewhere effect
#> 67                                      Loss aversion
#> 68                               Mere exposure effect
#> 69                                     Money illusion
#> 70                            Moral credential effect
#> 71                                    Negativity bias
#> 72                             Neglect of probability
#> 73                                      Normalcy bias
#> 74                                  Not invented here
#> 75                         Observer-expectancy effect
#> 76                                      Omission bias
#> 77                                      Optimism bias
#> 78                                     Ostrich effect
#> 79                                       Outcome bias
#> 80                              Overconfidence effect
#> 81                                         Pareidolia
#> 82                                   Pygmalion effect
#> 83                                     Pessimism bias
#> 84                                   Planning fallacy
#> 85                                       Present bias
#> 86                                Pro-innovation bias
#> 87                                    Projection bias
#> 88                             Pseudocertainty effect
#> 89                                          Reactance
#> 90                               Reactive devaluation
#> 91                                   Recency illusion
#> 92                                     Restraint bias
#> 93                             Rhyme as reason effect
#> 94                                  Risk compensation
#> 95                                      Salience bias
#> 96                                     Selection bias
#> 97                               Selective perception
#> 98                                  Semmelweis reflex
#> 99                         Sexual overperception bias
#> 100                                Singularity effect
#> 101                            Social comparison bias
#> 102                          Social desirability bias
#> 103                                   Status quo bias
#> 104                                      Stereotyping
#> 105                              Subadditivity effect
#> 106                             Subjective validation
#> 107                                       Surrogation
#> 108                                 Survivorship bias
#> 109                                  Time-saving bias
#> 110                               Third-person effect
#> 111                     Parkinson's law of triviality
#> 112                                 Weber–Fechner law
#> 113                        Well travelled road effect
#> 114                        Women are wonderful effect
#> 115                                    Zero-risk bias
#> 116                                     Zero-sum bias
#> 117                               Actor-observer bias
#> 118                                    Authority bias
#> 119                                Cheerleader effect
#> 120                  Defensive attribution hypothesis
#> 121                                   Egocentric bias
#> 122                         Extrinsic incentives bias
#> 123                            False consensus effect
#> 124                             False uniqueness bias
#> 125                     Fundamental attribution error
#> 126                           Group attribution error
#> 127                                       Halo effect
#> 128                    Illusion of asymmetric insight
#> 129                       Illusion of external agency
#> 130                          Illusion of transparency
#> 131                              Illusory superiority
#> 132                                      Ingroup bias
#> 133                             Just-world hypothesis
#> 134                                        Moral luck
#> 135                                    Naïve cynicism
#> 136                                     Naïve realism
#> 137                         Outgroup homogeneity bias
#> 138                                 Self-serving bias
#> 139                           Shared information bias
#> 140                              System justification
#> 141                             Trait ascription bias
#> 142                        Ultimate attribution error
#> 143                         Worse-than-average effect
#> 144                                Bizarreness effect
#> 145                                    Context effect
#> 146                                 Cross-race effect
#> 147                                      Cryptomnesia
#> 148                                   Egocentric bias
#> 149                                Fading affect bias
#> 150                                      False memory
#> 151                                 Generation effect
#> 152                                     Google effect
#> 153                           Leveling and sharpening
#> 154                       Levels-of-processing effect
#> 155                             Misinformation effect
#> 156                                   Modality effect
#> 157                        Mood-congruent memory bias
#> 158                           Part-list cueing effect
#> 159                                     Peak-end rule
#> 160                        Picture superiority effect
#> 161                 Socioemotional selectivity theory
#> 162                                    Primacy effect
#> 163                                    recency effect
#> 164                            serial position effect
#> 165                                 Reminiscence bump
#> 166                                  Source confusion
#> 167                                    Spacing effect
#> 168                                  Spotlight effect
#> 169                                    Suggestibility
#> 170                                      Tachypsychia
#> 171                                Telescoping effect
#> 172                                    Testing effect
#> 173                                 Tip of the tongue
#> 174                               von Restorff effect
#> 175                                  Zeigarnik effect
#> 176                                           General
#> 177                                            Memory

  # }